Change GlobalDefines.cs build
GlobalDefines.cs holds information like version number and license about Tasque. It is generated by a special MSBuild task ("SetDataAtBuildTime") from GlobalDefines.cs.in after the build of build.csproj (which builds the SetDataAtBuildTime MSBuild task). GlobalDefines.cs is then linked to by the various .csproj files of the solution which include it in their build.

* build.mdproj: has been renamed to build.csproj
* build.csproj:
	- removed all stuff related to MSBuild.Community.Tasks, which is not needed anymore
	- added build configuration to build SetDataAtBuildTime.cs

* tasque.sln:
	- reflect build proj changes

* .gitignore:
	- reflect removal of MSBuild.Community.Tasks
	- add ignore of .mdb files

* MSBuild.Community.Tasks.dll and MSBuild.Community.Tasks.Targets: Those are not needed anymore, since everything is now handled by a custom self-written task

* GlobalDefines.cs.in and GlobalDefines.cs:
	- include some more data about the project