Add a TreeModel implementation with independent backing store
The standard implementation of the Gtk.TreeModel interface, which is Gtk.TreeStore, is considered the model in the MVC oriented Gtk framework. As such it has its own internal means of storing data. In order to use an alternative backing store, e.g. a generic Collection, a custom TreeModel implementation has been built.

Resources used:
* mono docs
* gtk+ docs
* article: http://www.mono-project.com/ImplementingGInterfaces
* reference implementation: NodeStore.cs at https://github.com/mono/gtk-sharp/blob/master/gtk/NodeStore.cs