Remove last Gtk# dep from model: IBackend.GetPreferencesWidget ()
The method IBackend.GetPreferencesWidget () has been substituted by the read-only property IBackend.Preferences, which is of type IBackendPreferences. This interface has been created to provide an abstract handle for the class PreferencesDialog.cs in the application. A concrete preferences widget is being provided along with the backend implementation: DummyPreferences.cs.

This commit removes the last residuals of Gtk# in the model.