Outsource tray logic from Application.cs
* AppIndicator.cs: Tray icon for Ubuntu's Unity desktop. This
should fix the bug, that the tray icon isn't displayed in Unity. (Though
the fix is not yet complete: need to change build settings, to only use
this if appindicator-sharp is present.)
* StatusIcon.cs: Fallback systray for non-Unity desktops
* GtkTrayBase.cs: Common base class of the above two.
* TrayModelcs: The viewmodel that provides all the data for the
tray to work.
* Tasque.Gtk.csproj: IMPORTANT: Currently there is no check if
appinidicator-sharp is present on the build machine. This must be added
for the build to work on systems where appindicator is not available.

Note: AppIndicator tray and StatusIcon tray work a little different:
Toggling the main window in AppIndicator is done via a menu item
(due to a limitation of AppIndicator API),
whereas StatusIcon toggles via clicking the icon.