Implement Object Service feature
Sometimes a situation occurs where objects which are totally alike
in behavior and state need to be used by other objects and are
therefore instantiated by those other classes. It then is somewhat
cumbersome to instantiate these objects again and again although
nobody actually needs that many instances. One instance would suffice.
One approach to solve this would be to apply the Singleton pattern.
Doing that has big implications (see GoF book) and possibly a disturbing impact on future development. The singleton pattern
approach also doesn't fully match the problem: The objects in
question may not be single to the whole application but only to a
certain subtree of the UI tree structure.
To provide a solution to this I introduce "Object Service" with this commit: Descendants of a UI subtree ask their ancestors, if they can
provide them with an object of e specified type. Any object in the
tree can provide this service by registering objects to the Object
Service.