Fix AllCategory doesn't have elements.
AllCategory has been removed. There is now only an entry in the
ComboBox displaying the categories that says "All". This is the only
place in Tasque where the AllCategory was actually used, so it's no
big loss.

* Backend.cs: CreateTask doesn't throw anymore if no category is
provided, but uses the default category of the backend instead.

* TaskWindow.cs: The TreeModel of the categories' combobox is no
longer a model of Categories but of strings, of which the first item
is "All" and the rest is filled with the backend's categories' names.