Introduce Prefix build variable
The prefix variable is the same as in autotool's configure scripts. By default
it points to build/out/. The projects output ontop of prefix follows the
common Unix directory structure.

In favor of the prefix variable the DataDir variable has been dropped. The 
DataDir, which is the same as the Unix share directory, can be deduced as
Prefix/share.

The Prefix variable is accessible in code via the GlobalDefines class.

The SetDataAtBuildTime task has been updated to reflect the transition
to the Prefix variable.