Update translation compilation settings
* CompileTranslations.cs: MSBuild task to compile translations using
msgfmt.
* DeleteCompiledTranslations.cs: MSBuild task to clean the translation
compilation output.
* .gitignore: ignore translation compilation output in vcs

* translation.mdproj:
This proj file is now origanized in two parts. The above 56 lines are for the
integration in MD. Everything below is part of the MSBuild build
instructions.
The project outputs to $prefix/share/locale which is by default
build/out/share/locale. It uses to MSBuild tasks (described above) to
manage translation compilation. The respective build targets, the tasks
belong to, are "Build" and "Clean". The "Rebuild" target is a superposition
of "Clean" and "Build".

* libtasque.csproj: For an unkonwn reason a line in Options.cs is shown as
being marked for translation, though it's really not. Exclude it therefore.