[libtasque] Category: Make class abstract; Resolve DefaultCategory mess
libtasque:
* Category.cs: Child classes rely on getting only their flavor of Category.
Hence make category abstract and ensure that only instances of child
categories are around. This also means the ctor needs not be public.
* Backend.cs: Does not create a default category in ctor anymore. Declares
that member as abstract and lets the child class handle it. Provides
guidelines for implementation.
---------------------
DummyBackend:
The backend allows interaction with DefaultCategory member only after
initialization. WorkCategory is declared as DefaultCategory;
---------------------
SQLiteBackend:
Same as DummyBackend
Additionally: Changed namespace of pref widget class.